requirements.txt
flask>=3.0
face_recognition>=1.3.0
Pillow>=10.0
numpy>=1.24
